add config use_plain_encoding and move EncodingInfoResolver to ExecEnv
