+++ be/src/common/config.cpp
@@ -1600,6 +1600,8 @@ DEFINE_mBool(read_cluster_cache_opt_verbose_log, "false");
 
 DEFINE_mDouble(compression_min_space_saving, "0.1");
 
+DEFINE_Bool(use_plain_encoding, "true");
+
 // clang-format off
 #ifdef BE_TEST
 // test s3
