+++ fe/fe-core/src/main/java/org/apache/doris/catalog/TabletStatMgr.java
@@ -150,8 +150,6 @@ public class TabletStatMgr extends MasterDaemon {
                 Long tableTotalLocalSegmentSize = 0L;
                 Long tableTotalRemoteIndexSize = 0L;
                 Long tableTotalRemoteSegmentSize = 0L;
-                Long tableLocalCommonIndexSize = 0L;
-                Long tableRemoteCommonIndexSize = 0L;
 
                 Long tableRemoteDataSize = 0L;
 
@@ -219,8 +217,6 @@ public class TabletStatMgr extends MasterDaemon {
                                     tableTotalLocalSegmentSize += replica.getLocalSegmentSize();
                                     tableTotalRemoteIndexSize += replica.getRemoteInvertedIndexSize();
                                     tableTotalRemoteSegmentSize += replica.getRemoteSegmentSize();
-                                    tableLocalCommonIndexSize += replica.getLocalCommonIndexSize();
-                                    tableRemoteCommonIndexSize += replica.getRemoteCommonIndexSize();
                                 }
 
                                 tableDataSize += tabletDataSize;
@@ -267,8 +263,7 @@ public class TabletStatMgr extends MasterDaemon {
                             tableDataSize, tableTotalReplicaDataSize,
                             tableRemoteDataSize, tableReplicaCount, tableRowCount, 0L, 0L,
                             tableTotalLocalIndexSize, tableTotalLocalSegmentSize,
-                            tableTotalRemoteIndexSize, tableTotalRemoteSegmentSize, tableLocalCommonIndexSize,
-                            tableRemoteCommonIndexSize));
+                            tableTotalRemoteIndexSize, tableTotalRemoteSegmentSize));
 
                     if (LOG.isDebugEnabled()) {
                         LOG.debug("finished to set row num for table: {} in database: {}",
